run seeds at start if database is empty

+++ src/database/seeds/01_call_price.js
@@ -1,15 +1,22 @@
-exports.seed = function (knex) {
-  return knex.schema.raw(`
-    delete from call_price;
-  `).raw(`
-    insert into call_price
-      (origin, destination, cents_per_minute)
-    values
-      ('011', '016', 190),
-      ('016', '011', 290),
-      ('011', '017', 190),
-      ('017', '011', 270),
-      ('011', '018', 090),
-      ('018', '011', 190);
+exports.seed = async function (knex) {
+  let callPrices = await knex.schema.raw(`
+    select * from call_price
   `);
+  if (callPrices.rowCount == 0) {
+    return knex.schema.raw(`
+      delete from call_price;
+    `).raw(`
+      insert into call_price
+        (origin, destination, cents_per_minute)
+      values
+        ('011', '016', 190),
+        ('016', '011', 290),
+        ('011', '017', 190),
+        ('017', '011', 270),
+        ('011', '018', 090),
+        ('018', '011', 190);
+    `);
+  }
+
+  return;
 };
